fix: melhorando métodos adicionando paginação

+++ src/main/java/com/java/boilerplate/repository/IChatMessagesRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 @Repository
 public interface IChatMessagesRepository extends IBaseRepository<ChatMessages> {
-    List<ChatMessages> findConversation(@Param("myId") Long myId, @Param("contactId") Long contactId, @Param("nextEntryId") Long nextEntryId, Pageable pageable );
+    List<ChatMessages> findConversation(@Param("myId") Long myId, @Param("contactId") Long contactId, @Param("nextEntryId") Integer nextEntryId, Pageable pageable );
     List<ChatMessages> findLastMessageExchange(@Param("user") Users user, @Param("contact") Users contact, Pageable pageable);
     Long countUnreadMessages(@Param("contactId") Long contactId, @Param("myId") Long myId);
 }
